#194 Create services for each entity, some improvements in stock style
